add resource and Auth to social media crd

+++ app/Http/Controllers/Api/SocialMediaController.php
@@ -7,52 +7,50 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use App\Traits\ResponseJson;
 use Illuminate\Http\Request;
+use App\Http\Resources\socialMediaResource ;
 use App\Models\SocialMedia;
 
 class SocialMediaController extends Controller
 {
     use ResponseJson;
-    
-    public function index()
-    {
-        $socialMedia = SocialMedia::all();
-        if($socialMedia){
-            return $this->jsonResponseWithoutMessage($socialMedia, 'data',200);
-        }
-    }
 
     public function create(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'user_id' => 'required',
-            'facebook' => 'required_without_all:twitter,instagram',
-            'twitter' => 'required_without_all:facebook,instagram',
-            'instagram' => 'required_without_all:facebook,twitter',
-        ]);
+        $userExists = SocialMedia::where('user_id',Auth::id())->count();
+        if($userExists == 0){
+            $validator = Validator::make($request->all(), [
+                'facebook' => 'required_without_all:twitter,instagram',
+                'twitter' => 'required_without_all:facebook,instagram',
+                'instagram' => 'required_without_all:facebook,twitter',
+            ]);
 
-        if ($validator->fails()) {
-            return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
-        } 
-
-        SocialMedia::create($request->all());
-        return $this->jsonResponseWithoutMessage("Your Accounts Are Added Successfully", 'data', 200);
-        
+            if ($validator->fails()) {
+                return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
+            } 
+            $request['user_id'] = Auth::id();
+            SocialMedia::create($request->all());
+            return $this->jsonResponseWithoutMessage("Your Accounts Are Added Successfully", 'data', 200);
+        } else {
+            return $this->jsonResponseWithoutMessage("You can't Add Anothe Social Media Accounts", 'data', 500);
+        }
     }
 
     
     public function show(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'socialMedia_id' => 'required',
+            'user_id' => 'required',
         ]);
 
         if ($validator->fails()) {
             return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
         }
-
-        $socialMedia = SocialMedia::find($request->socialMedia_id);
-        if($socialMedia){
-            return $this->jsonResponseWithoutMessage($socialMedia, 'data',200);
+        if($request->user_id == Auth::id()){
+            $socialMedia = SocialMedia::where('user_id',Auth::id())->first();
+            if($socialMedia){
+                $socialMedia = new socialMediaResource($socialMedia);
+                return $this->jsonResponseWithoutMessage($socialMedia, 'data',200);
+            }
         }
     }
 
@@ -60,34 +58,35 @@ class SocialMediaController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'user_id' => 'required',
-            'socialMedia_id' => 'required',
             'facebook' => 'required_without_all:twitter,instagram',
             'twitter' => 'required_without_all:facebook,instagram',
             'instagram' => 'required_without_all:facebook,twitter',
 
         ]);
-
         if ($validator->fails()) {
             return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
         }
-
-        $socialMedia = SocialMedia::find($request->socialMedia_id);
-        $socialMedia->update($request->all());
-        return $this->jsonResponseWithoutMessage("Your Accounts Are Updated Successfully", 'data', 200);
+    
+        if($request->user_id == Auth::id()){
+            $socialMedia = SocialMedia::where('user_id',Auth::id())->first();
+            $socialMedia->update($request->all());
+            return $this->jsonResponseWithoutMessage("Your Accounts Are Updated Successfully", 'data', 200);
+        }
     }
 
     public function delete(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'socialMedia_id' => 'required',
+            'user_id' => 'required',
         ]);
 
         if ($validator->fails()) {
             return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
         }
-
-        $socialMedia = SocialMedia::find($request->socialMedia_id);
-        $socialMedia->delete();
-        return $this->jsonResponseWithoutMessage("Your Accounts Are Deleted Successfully", 'data', 200);
+        if($request->user_id == Auth::id()){
+            $socialMedia = SocialMedia::where('user_id',Auth::id())->first();
+            $socialMedia->delete();
+            return $this->jsonResponseWithoutMessage("Your Accounts Are Deleted Successfully", 'data', 200);
+        }
     }
 }
